Add an example to Resolving Third Party Library
As I started with Liferay development, I didn't understand well, how to resolve third party libraries. As it is written here, there are many different ways, But couldn't understand how they exactly works. Only after a long try and error period. And resolving axis-2, which is a pain in the ass. I understand the simplicity of it. Therefore I crated this exampel to help other juniors to understand the process involved to it. I would be honored if the link stays and i might help some people with their strugle.